Working on Heart Controller
In the midst of rewriting the heartController so that it first checks whether or not a user has already marked a particular comment as a favorite. If so, then we go ahead and DESTROY the favorite. If the user has not yet marked this particular comment as a favorite, then we go ahead and CREATE a new favorite. However, this is currently breaking our commentRouter and I'm not sure why.